Revert "using Long instead of Integer for the buffers"
This reverts commit 1140a0209aae84137cba73621d4f7f4e5dc096db.

+++ src/main/java/org/geppetto/solver/sph/PCISPHCheckPoint.java
@@ -39,11 +39,11 @@ import java.util.List;
 public class PCISPHCheckPoint {
 	
 	public List<Float> acceleration = new ArrayList<Float>();
-	public List<Long> gridCellIndex = new ArrayList<Long>();
-	public List<Long> gridCellIndexFixedUp = new ArrayList<Long>();
+	public List<Integer> gridCellIndex = new ArrayList<Integer>();
+	public List<Integer> gridCellIndexFixedUp = new ArrayList<Integer>();
 	public List<Float> neighborMap = new ArrayList<Float>();
-	public List<Long> particleIndex = new ArrayList<Long>();
-	public List<Long> particleIndexBack = new ArrayList<Long>();
+	public List<Integer> particleIndex = new ArrayList<Integer>();
+	public List<Integer> particleIndexBack = new ArrayList<Integer>();
 	public List<Float> position = new ArrayList<Float>();
 	public List<Float> pressure = new ArrayList<Float>();
 	public List<Float> rho = new ArrayList<Float>();
